% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{davis_data}
\alias{davis_data}
\title{Davis Friendship Network}
\format{
A data frame with 89 rows and 2 variables.
}
\source{
davis_25.csv
}
\usage{
davis_data
}
\description{
Friendship relations in a small group.
}
\examples{
data(davis_data)
head(davis_data)
}
\keyword{datasets}
