% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{drugs_data}
\alias{drugs_data}
\title{Drug Deaths Data}
\format{
A data frame with 11981 rows and 48 variables.
}
\source{
DrugsDeaths_5.csv
}
\usage{
drugs_data
}
\description{
Statistics on drug-related deaths across regions.
}
\examples{
data(drugs_data)
summary(drugs_data)
}
\keyword{datasets}
