% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{florentine_data}
\alias{florentine_data}
\title{Florentine Families Network}
\format{
A data frame with 16 rows and 16 variables.
}
\source{
florentineFamily_2.csv
}
\usage{
florentine_data
}
\description{
Relationships among Renaissance Florentine families.
}
\examples{
data(florentine_data)
florentine_data
}
\keyword{datasets}
