% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{florentine_party}
\alias{florentine_party}
\title{Florentine Party Invitations}
\format{
A data frame with 16 rows and 3 variables.
}
\source{
party_9.csv
}
\usage{
florentine_party
}
\description{
Party invitation network among Florentine elite.
}
\examples{
data(florentine_party)
florentine_party
}
\keyword{datasets}
