% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{HSE_one_love}
\alias{HSE_one_love}
\title{HSE One Love Survey}
\format{
A data frame with 450 rows and 3 variables.
}
\source{
HSE_one_love_23.csv
}
\usage{
HSE_one_love
}
\description{
Survey of students on topics of love and relationships.
}
\examples{
data(HSE_one_love)
head(HSE_one_love)
}
\keyword{datasets}
