% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{imdb_1000}
\alias{imdb_1000}
\title{IMDb Top 1000 Movies}
\format{
A data frame with 979 rows and 6 variables.
}
\source{
imdb_1000_23.csv
}
\usage{
imdb_1000
}
\description{
Metadata on top 1000 movies from IMDb.
}
\examples{
data(imdb_1000)
head(imdb_1000)
}
\keyword{datasets}
