% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{Jane_Eyre}
\alias{Jane_Eyre}
\title{Jane Eyre Text}
\format{
A data frame with 21005 rows and 2 variables.
}
\source{
Jane_Eyre_13.csv
}
\usage{
Jane_Eyre
}
\description{
Textual dataset from the novel "Jane Eyre".
}
\examples{
data(Jane_Eyre)
head(Jane_Eyre)
}
\keyword{datasets}
