% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{migration_data}
\alias{migration_data}
\title{Migration Network Data}
\format{
A data frame with 10 rows and 10 variables.
}
\source{
AdjacencyDirectedWeighted_14.csv
}
\usage{
migration_data
}
\description{
Directed and weighted adjacency matrix of international migration flows.
}
\examples{
data(migration_data)
head(migration_data)
}
\keyword{datasets}
