% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{netflix_data}
\alias{netflix_data}
\title{Netflix Ratings Data}
\format{
A data frame with 1000 rows and 7 variables.
}
\source{
netflix.xlsx
}
\usage{
netflix_data
}
\description{
Netflix user rating dataset.
}
\examples{
data(netflix_data)
str(netflix_data)
}
\keyword{datasets}
