% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{netflix_titles}
\alias{netflix_titles}
\title{Netflix Titles Metadata}
\format{
A data frame with 6234 rows and 12 variables.
}
\source{
netflix_titles.csv
}
\usage{
netflix_titles
}
\description{
Information about Netflix titles (movies and shows).
}
\examples{
data(netflix_titles)
head(netflix_titles)
}
\keyword{datasets}
