% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{newman1}
\alias{newman1}
\title{Newman Network 1}
\format{
A data frame with 58594 rows and 2 variables.
}
\source{
newman1_25.txt
}
\usage{
newman1
}
\description{
Network data used in community detection.
}
\examples{
data(newman1)
head(newman1)
}
\keyword{datasets}
