% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{newman2}
\alias{newman2}
\title{Newman Network 2}
\format{
A data frame with 16725 rows and 3 variables.
}
\source{
newman2_25.txt
}
\usage{
newman2
}
\description{
Second network for testing graph algorithms.
}
\examples{
data(newman2)
head(newman2)
}
\keyword{datasets}
