% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{predimed_data}
\alias{predimed_data}
\title{Predimed Nutrition Study}
\format{
A data frame with 6324 rows and 15 variables.
}
\source{
predimed_18.csv
}
\usage{
predimed_data
}
\description{
Data from a nutritional intervention trial (PREDIMED).
}
\examples{
data(predimed_data)
summary(predimed_data)
}
\keyword{datasets}
