% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{tales_data}
\alias{tales_data}
\title{Tales Tutorial Data}
\format{
A data frame with 10 rows and 1 variable.
}
\source{
tales_tutorial_3.csv
}
\usage{
tales_data
}
\description{
Simple dataset used for demonstration purposes.
}
\examples{
data(tales_data)
tales_data
}
\keyword{datasets}
