% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{terrorism_data}
\alias{terrorism_data}
\title{Terrorism Dataset}
\format{
A data frame with 181691 rows and 12 variables.
}
\source{
data_7.csv
}
\usage{
terrorism_data
}
\description{
Global terrorism incidents data.
}
\examples{
data(terrorism_data)
summary(terrorism_data)
}
\keyword{datasets}
